requirements.txt
streamlit
PyPDF2
python-dotenv
requests
supabase
groq
gtts
pillow
numpy
opencv-python
ffmpeg-python==0.2.0